requirements.txt
certifi==2018.1.18
diff-match-patch==20121119
Django==1.10.5
django-crispy-forms==1.6.1
django-debug-toolbar==1.9.1
django-formtools==2.0
django-import-export==1.0.0
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.10.3
jdcal==1.3
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.5.1
Pillow==4.0.0
PyMySQL==0.8.0
PyYAML==3.12
six==1.11.0
sqlparse==0.2.4
tablib==0.11.5
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
